requirements.txt
langchain
langchain-chroma
langchain-community
langchain-openai
python-dotenv
chromadb